Logistic ridge regression.

Fits a logistic ridge regression model. Optionally, the ridge regression parameter is chosen automatically using the method proposed by Cule et al (2012).

logisticRidge(formula, data, lambda = "automatic", nPCs = NULL,
scaling = c("corrForm", "scale", "none"), ...)

# S3 method for ridgeLogistic
coef(object, all.coef = FALSE, ...)

# S3 method for ridgeLogistic
plot(x, y = NULL, ...)

# S3 method for ridgeLogistic
predict(object, newdata = NULL, type = c("link", "response"), 
    na.action = na.pass, all.coef = FALSE, ...)

# S3 method for ridgeLogistic
print(x, all.coef = FALSE, ...)

# S3 method for ridgeLogistic
summary(object, all.coef = FALSE, ...)

# S3 method for summary.ridgeLogistic
print(x, digits = max(3, getOption("digits") - 3),
signif.stars = getOption("show.signif.stars"), ...)

Arguments

formula

a formula expression as for regression models, of the form response ~ predictors. See the documentation of formula for other details.

data

an optional data frame in which to interpret the variables occuring in formula.

lambda

A ridge regression parameter. If lambda is "automatic" (the default), then the ridge parameter is chosen automatically using the method of Cule et al (2012).

nPCs

The number of principal components to use to choose the ridge regression parameter, following the method of Cule et al (2012). It is not possible to specify both lambda and nPCs.

scaling

The method to be used to scale the predictors. One of "corrform"(the default) scales the predictors to correlation form, such that the correlation matrix has unit diagonal. "scale"Standardizes the predictors to have mean zero and unit variance. "none"No scaling.

object

A ridgeLogistic object, typically generated by a call to linearRidge.

newdata

An optional data frame in which to look for variables with which to predict. If omitted, the fitted values are used.

type

the type of prediction required. The default predictions are of log-odds (probabilities on logit scale) and type = "response" gives the predicted probabilities.

na.action

function determining what should be done with missing values in newdata. The default is to predict NA.

all.coef

Logical. Should results be returned for all ridge regression penalty parameters (all.coef = TRUE), or only for the ridge parameter chosen automatically using the method of Cule et al?

x

An object of class ridgeLogistic (for the print.ridgeLogistic and plot.ridgeLogistic functions) or an object of class summary.ridgeLogistic (for the print.summary.ridgeLogistic function)

y

Dummy argument for compatibility with the default plot method. Ignored.

digits

minimum number of significant digits to be used for most numbers

signif.stars

logical; if TRUE, P-values are additionally encoded visually as significance stars in order to help scanning of long coefficient tables. It defaults to the show.signif.stars slot of options.

...

Additional arguments to be passed to or from other methods.

Details

If an intercept is present in the model, its coefficient is not penalised. If you want to penalise an intercept, put in your own constant term and remove the intercept.

Value

An object of class "ridgeLogistic", with components:

automatic

Was lambda chosen automatically?

call

The matched call.

coef

A named vector of fitted coefficients.

df

A vector of degrees of freedom of the model fit and degrees of freedom for variance.

Inter

Was in antercept included?

isScaled

Were the predictors scaled before the model was fitted?

lambda

The ridge regression parameter.

scales

The scales used to standardize the predictors.

terms

The terms object used.

x

The scaled predictor matrix.

xm

A vector of means of the predictors.

y

The response.

And optionally the component

nPCs

The number of principal components used to compute the ridge regression parameter.

References

A semi-automatic method to guide the choice of ridge parameter in ridge regression. Cule, E. and De Iorio, M. (2012) arXiv:1205.0686v1 [stat.AP]

Author

Erika Cule

See also

Examples

data(GenBin)
mod <- logisticRidge(Phenotypes ~ ., data = as.data.frame(GenBin))
summary(mod)
#> 
#> Call:
#> logisticRidge(formula = Phenotypes ~ ., data = as.data.frame(GenBin))
#> 
#> 
#> Coefficients:
#>              Estimate Scaled estimate Std. Error (scaled) t value (scaled)
#> (Intercept)  0.002782              NA                  NA               NA
#> SNP1         0.166067        1.914369            0.666082            2.874
#> SNP2         0.110745        0.269635            0.963936            0.280
#> SNP3        -0.284802       -0.491809            0.999027           -0.492
#> SNP4        -1.007615       -1.422131            0.990998           -1.435
#> SNP5         0.911082        0.910170            0.986458            0.923
#> SNP6        -0.026877       -0.355500            0.959424           -0.371
#> SNP7        -0.068540       -1.022005            0.922686           -1.108
#> SNP8         0.149821        1.729257            0.667675            2.590
#> SNP9        -0.159520       -0.795170            0.948859           -0.838
#> SNP10       -0.207470       -1.214400            0.893062           -1.360
#> SNP11        0.027842        0.067788            0.999575            0.068
#> SNP12        0.090563        0.651787            0.958376            0.680
#> SNP13       -0.051310       -0.705257            0.947386           -0.744
#> SNP14       -0.196030       -0.990720            0.904846           -1.095
#>             Pr(>|t|)   
#> (Intercept)       NA   
#> SNP1         0.00405 **
#> SNP2         0.77969   
#> SNP3         0.62252   
#> SNP4         0.15127   
#> SNP5         0.35618   
#> SNP6         0.71098   
#> SNP7         0.26802   
#> SNP8         0.00960 **
#> SNP9         0.40202   
#> SNP10        0.17389   
#> SNP11        0.94593   
#> SNP12        0.49644   
#> SNP13        0.45662   
#> SNP14        0.27356   
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Ridge paramter: 0.1244006, chosen automatically, computed using 3 PCs
#> 
#> Degrees of freedom: model 7.02 , variance 4.066 
#>