Internal functions for logistic ridge regression.

Internal functions for logisitc ridge regression.

computeRidgeLogistic(X, y, k, intercept = TRUE, doff = FALSE)
updateBeta(B, X, y, k, intercept = TRUE, doff = FALSE) 
objectiveFunction(B, X, y, k, intercept = TRUE)

Arguments

X

Matrix of predictors.

y

vector of outcomes.

k

ridge regression parameter.

intercept

does the model have an intercept?

doff

should degrees of freedom of the model be computed?

Details

These functions are called in the function logisticRidge. They are not for calling directly by the user.

Value

computeRidgeLogistic returns the fitted logistic ridge regression coefficients. If doff = TRUE it also returns the degrees of freedom of the model and the degrees of freedom for variance. updateBeta returns the fitted coefficients after one iteration of the Newton-Raphson algorithm. If doff = TRUE, it also returns the penalty matrix and weights matrix used to compute the degrees of freedom. objectiveFunction returns the objective function for the current iteration of the Newton-Raphson algorithm.

References

A semi-automatic method to guide the choice of ridge parameter in ridge regression. Cule, E. and De Iorio, M. (2012) arXiv:1205.0686v1 [stat.AP]

Author

Erika Cule

Note

These functions are not to be called directly by the user. They should be called via logisticRidge.

See also